♻️ refactor(home): extract inbox section assembly into a hook
Move HomeInbox's data fetching and section-assembly logic into useHomeInboxSections; index.tsx now only does render dispatch based on the hook's status/sections. Pure refactor, no behavior change.
